requirements.txt
moviepy==1.0.3
openai==1.9.0
elevenlabs==0.2.27
PyYAML==6.0.1
requests==2.31.0
python-dotenv==1.0.1
# These bottom two lines are to avoid deprecation 
Pillow==9.5.0
opencv-python==4.9.0.80
gradio==4.19.1
pytest==8.0.1
pytest-asyncio==0.23.5